todo: add RESTful endpoint to fetch user data
