+++ pages/index.html
@@ -6,7 +6,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Home</title>
-    <script src="/assets/src/index.js" defer></script>
+    <link rel="stylesheet" href="/assets/styles/index.css">
+    <script type="module" src="/assets/src/index.js" defer></script>
 </head>
 
 <body>
